+++ mqtt/mqttd/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 [dependencies]
 anyhow = "1.0"
 async-trait = "0.1"
+cfg-if = "1.0"
 chrono = "0.4"
 clap = "2.33"
 futures-util = { version = "0.3", features = ["sink"] }
