+++ mqtt/mqttd/src/lib.rs
@@ -11,5 +11,5 @@
     clippy::missing_errors_doc
 )]
 
-pub mod broker;
+pub mod app;
 pub mod tracing;
